Pulled up r4748 from trunk.
git-svn-id: https://svn.calendarserver.org/repository/calendarserver/CalendarServer/branches/release/CalendarServer-2.4-dev@4769 e27351fd-9f3e-4f54-a53b-843176b1656c

+++ twistedcaldav/upgrade.py
@@ -1,3 +1,4 @@
+# -*- test-case-name: twistedcaldav.test.test_upgrade -*-
 ##
 # Copyright (c) 2008 Apple Inc. All rights reserved.
 #
@@ -177,6 +178,10 @@ def upgrade_to_1(config):
         try:
             for cal in os.listdir(homePath):
                 calPath = os.path.join(homePath, cal)
+                if not os.path.isdir(calPath):
+                    # Skip non-directories; these might have been uploaded by a
+                    # random DAV client, they can't be calendar collections.
+                    continue
                 if cal == 'notifications':
                     # Delete the old, now obsolete, notifications directory.
                     rmdir(calPath)
